[FP-59] Refactor: 체험 상페 페이지 리뷰 별점 소숫점 1자리까지만 나오게 수정
